requirements/base.txt
pycryptodome==3.9.7
python-dotenv==0.12.0
Flask==1.1.1
Flask-CORS==3.0.8
flask-restx==0.2.0